Add files via upload
These are datasets that were published as part of a 2017 study at Cardiff University. The study was concerned with finding a discreet way to monitor mood for mental health purposes using human interaction with smartphones. These datasets and the rest of the documentation can be found at: https://reshare.ukdataservice.ac.uk/852732/